채팅기능 비로그인 유저 예외처리

+++ src/main/java/com/example/shortform/controller/ChatRoomController.java
@@ -5,6 +5,7 @@ import com.example.shortform.dto.request.ChatRoomRequestDto;
 import com.example.shortform.dto.resonse.ChatMessageListDto;
 import com.example.shortform.dto.resonse.ChatRoomListResponseDto;
 import com.example.shortform.dto.resonse.ChatRoomResponseDto;
+import com.example.shortform.exception.UnauthorizedException;
 import com.example.shortform.service.ChatRoomService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -22,29 +23,43 @@ public class ChatRoomController {
     @PostMapping("/chat/rooms")
     public HashMap<String, Object> createChatRoom(@RequestBody ChatRoomRequestDto requestDto,
                                   @AuthenticationPrincipal PrincipalDetails principalDetails) {
-        chatRoomService.createChatRoom(requestDto, principalDetails);
-        HashMap<String, Object> result = new HashMap<>();
-        result.put("result", "true");
-        return result;
+        if (principalDetails != null) {
+            chatRoomService.createChatRoom(requestDto, principalDetails);
+            HashMap<String, Object> result = new HashMap<>();
+            result.put("result", "true");
+            return result;
+        } else {
+            throw new UnauthorizedException("로그인 후 이용가능합니다.");
+        }
     }
 
     @GetMapping("/chat/rooms")
     public List<ChatRoomListResponseDto> getAllMyRooms(@AuthenticationPrincipal PrincipalDetails principalDetails) throws ParseException {
-        return chatRoomService.getAllMyRooms(principalDetails);
+        if (principalDetails != null)
+            return chatRoomService.getAllMyRooms(principalDetails);
+        else
+            throw new UnauthorizedException("로그인 후 이용가능합니다.");
     }
 
     @GetMapping("/chat/rooms/{roomId}/messages")
     public ChatMessageListDto getAllMessages(@PathVariable Long roomId,
                                              @AuthenticationPrincipal PrincipalDetails principalDetails) {
-        return chatRoomService.getAllMessages(roomId, principalDetails);
+        if (principalDetails != null)
+            return chatRoomService.getAllMessages(roomId, principalDetails);
+        else
+            throw new UnauthorizedException("로그인 후 이용가능합니다.");
     }
 
     @DeleteMapping("/chat/rooms/{roomId}")
     public HashMap<String, Object> deleteRoom(@PathVariable Long roomId,
                                               @AuthenticationPrincipal PrincipalDetails principalDetails) {
-        chatRoomService.deleteRoom(roomId, principalDetails);
-        HashMap<String, Object> result = new HashMap<>();
-        result.put("result", "true");
-        return result;
+        if (principalDetails != null) {
+            chatRoomService.deleteRoom(roomId, principalDetails);
+            HashMap<String, Object> result = new HashMap<>();
+            result.put("result", "true");
+            return result;
+        } else {
+            throw new UnauthorizedException("로그인 후 이용가능합니다.");
+        }
     }
 }
